types: extract thunk types to thunks.ts
